requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.9.14
dearpygui==1.7.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
Pillow==9.2.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.3.2.post1
toml==0.10.2
